requirements.txt
requests
python-dotenv
puzpy